+++ .eslintrc.js
@@ -17,5 +17,9 @@ module.exports = {
     'no-continue': 'off',
     'no-await-in-loop': 'warn',
     'no-nested-ternary': 'off',
+    '@typescript-eslint/no-loop-func': 'off',
+    'import/prefer-default-export': 'off',
+    '@typescript-eslint/comma-dangle': 'off',
+    'require-await': 'off',
   },
 };