+++ api/src/routes/index.ts
@@ -1,6 +1,8 @@
 import { Router } from 'express';
 
 import ai from './ai';
+import datasetItems from './dataset-items';
+import datasets from './datasets';
 import folders from './folders';
 import importRouter from './import';
 import sdk from './sdk';
@@ -15,5 +17,7 @@ router.use('/sdk', sdk);
 router.use('/tags', tags);
 router.use('/folders', folders);
 router.use('/import', importRouter);
+router.use('/datasets', datasets);
+router.use('/datasets', datasetItems);
 
 export default router;